move `Container` to the top level of `layout`
