+++ src/core/components/layout/index.tsx
@@ -1,5 +1,5 @@
 export { Columns, Column } from './components/columns/columns';
-export { Container } from './components/container/container';
+export { Container, ContainerProps } from './Container/Container';
 export { Hide } from './components/hide/hide';
 export { Stack } from './components/stack/stack';
 export { Tiles } from './components/tiles/tiles';
